+++ scripts/migrate.sh
@@ -2,5 +2,6 @@
 set -eu
 
 python manage.py migrate --noinput
+python manage.py createcachetable
 python manage.py bootstrap_project
 python manage.py grant_database_access